requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
numpy==1.24.3
psycopg2-binary==2.9.7
pycryptodome==3.19.0
scikit-learn==1.3.0
scikit-image==0.21.0
matplotlib==3.7.2
python-dotenv==1.0.0
Werkzeug==2.3.7